+++ monkey/common/agent_configuration/default_agent_configuration.py
@@ -35,6 +35,7 @@ CREDENTIAL_COLLECTOR_CONFIGURATION = tuple(
 RANSOMWARE_OPTIONS = {
     "encryption": {
         "enabled": True,
+        "file_extension": ".m0nk3y",
         "directories": {"linux_target_dir": "", "windows_target_dir": ""},
     },
     "other_behaviors": {"readme": True},