+++ monkey/monkey_island/cc/ui/src/services/configuration/ransomware.js
@@ -1,24 +1,32 @@
 const RANSOMWARE_SCHEMA = {
-   'title': 'Payloads',
+    'title': 'Payloads',
     'properties': {
         'encryption': {
             'title': 'Ransomware simulation',
             'type': 'object',
             'description': 'To simulate ransomware encryption, you\'ll need to provide Infection ' +
-            'Monkey with files that it can safely encrypt. On each machine where you would like ' +
-            'the ransomware simulation to run, create a directory and put some files in it.' +
-            '\n\nProvide the path to the directory that was created on each machine.',
+                'Monkey with files that it can safely encrypt. On each machine where you would like ' +
+                'the ransomware simulation to run, create a directory and put some files in it.' +
+                '\n\nProvide the path to the directory that was created on each machine.',
             'properties': {
                 'enabled': {
                     'title': 'Encrypt files',
                     'type': 'boolean',
                     'default': true,
                     'description': 'Ransomware encryption will be simulated by flipping every bit ' +
-                    'in the files contained within the target directories.'
+                        'in the files contained within the target directories.'
                 },
                 'info_box': {
                     'info': 'No files will be encrypted if a directory is not specified or doesn\'t ' +
-                    'exist on a victim machine.'
+                        'exist on a victim machine.'
+                },
+                'file_extension': {
+                    'title': 'File extension',
+                    'type': 'string',
+                    'format': 'valid-file-extension',
+                    'default': '.m0nk3y',
+                    'description': 'The file extension that the Infection Monkey will use for the ' +
+                        'encrypted file.'
                 },
                 'directories': {
                     'title': 'Directories to encrypt',
@@ -30,8 +38,8 @@ const RANSOMWARE_SCHEMA = {
                             'format': 'valid-ransomware-target-path-linux',
                             'default': '',
                             'description': 'A path to a directory on Linux systems that contains ' +
-                            'files that you will allow Infection Monkey to encrypt. If no ' +
-                            'directory is specified, no files will be encrypted.'
+                                'files that you will allow Infection Monkey to encrypt. If no ' +
+                                'directory is specified, no files will be encrypted.'
                         },
                         'windows_target_dir': {
                             'title': 'Windows target directory',
@@ -39,8 +47,8 @@ const RANSOMWARE_SCHEMA = {
                             'format': 'valid-ransomware-target-path-windows',
                             'default': '',
                             'description': 'A path to a directory on Windows systems that contains ' +
-                            'files that you will allow Infection Monkey to encrypt. If no ' +
-                            'directory is specified, no files will be encrypted.'
+                                'files that you will allow Infection Monkey to encrypt. If no ' +
+                                'directory is specified, no files will be encrypted.'
                         }
                     }
                 },